requirements.txt
# Core Dependencies
python-telegram-bot==20.7
gspread==5.12.0
oauth2client==4.1.3

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Async Support
aiohttp==3.9.1

# Performance
psutil==5.9.6

# Note: sqlite3, asyncio, datetime, logging, json, os are built-in Python modules
